requirements.txt
APScheduler==3.9.1
homie-helpers~=0.0.6
service-bootstrap~=0.0.1
requests~=2.27.1
